Move the changes to a new sniff.
